facilitacme_django/backend/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.4,<3.13
xlsxwriter>=1.3.7,<1.4
fpdf>=1.7.2,<1.8
gunicorn>=20.1.0,<20.2